requirements.txt
# VIGILANT.IN — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Scraping ──────────────────────────────────────────────────────────────
scrapy==2.11.0              # Spider framework for structured site scraping
playwright==1.42.0          # Browser automation for JS-heavy portals (GeM, PFMS)
beautifulsoup4==4.12.3       # HTML parsing
lxml==5.1.0                 # Fast XML/HTML parser
requests==2.31.0            # HTTP requests
requests-html==0.10.0       # Requests with JS rendering
httpx==0.27.0               # Async HTTP client

# ── PDF Processing ────────────────────────────────────────────────────────────
pdfplumber==0.10.3          # EC affidavit PDF extraction (best for tables)
PyMuPDF==1.24.0             # Fallback PDF text extraction (fitz)
camelot-py==0.11.0          # Table extraction from PDFs
pdfminer.six==20221105      # Low-level PDF text extraction

# ── NLP & Entity Resolution ───────────────────────────────────────────────────
spacy==3.7.4                # NER for names, monetary values, dates
# Run after install: python -m spacy download en_core_web_lg
dedupe==2.0.22              # Probabilistic record linkage / deduplication
fuzzywuzzy==0.18.0          # Fuzzy string matching
python-Levenshtein==0.25.1  # Fast Levenshtein distance (fuzzywuzzy dependency)
nameparser==1.1.3           # Human name parsing and normalization
indic-nlp-library==0.91     # Hindi/regional name transliteration

# ── Databases ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9      # PostgreSQL adapter
sqlalchemy==2.0.29          # ORM and query builder
alembic==1.13.1             # Database migrations
neo4j==5.19.0               # Neo4j graph database driver
redis==5.0.4                # Redis client (caching, job queue)

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==2.2.1               # Data manipulation and analysis
numpy==1.26.4               # Numerical computing
scipy==1.13.0               # Statistical functions (correlation)
python-dateutil==2.9.0      # Date parsing

# ── Machine Learning ──────────────────────────────────────────────────────────
xgboost==2.0.3              # Gradient boosting for risk scoring
scikit-learn==1.4.1         # ML utilities, preprocessing
joblib==1.4.0               # Model serialization

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.110.1            # REST API framework
uvicorn==0.29.0             # ASGI server
pydantic==2.7.0             # Data validation and serialization
python-multipart==0.0.9     # File upload support

# ── Workflow Orchestration ────────────────────────────────────────────────────
apache-airflow==2.9.0       # DAG scheduling and monitoring
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-redis==3.6.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1        # .env file loading
tenacity==8.2.3             # Retry logic for scraper failures
loguru==0.7.2               # Structured logging
rich==13.7.1                # Beautiful terminal output
click==8.1.7                # CLI argument parsing
pyyaml==6.0.1               # YAML config files
schedule==1.2.1             # Simple job scheduling (dev mode)
tqdm==4.66.2                # Progress bars

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.1.1               # Test framework
pytest-asyncio==0.23.6      # Async test support
pytest-cov==5.0.0           # Test coverage
responses==0.25.0           # Mock HTTP responses for tests
factory-boy==3.3.0          # Test data factories
